[WIP] Lift var declarations; begin supporting ops with Any
why this is WIP: it breaks the integration tests because by assigning
undefined to the vars, we end up forcing them to have type Any. hopefully when
we have type inference, we can do some fanciness around undefined so this
won't always happen, but until then since we have to anyway i expanded
support for binary ops between Anys to allow the integration tests to pass
again. however i am going to have to make a pretty significant change to how we
deal with runtime system functions so i'm committing