Create global object in notwasm (completes new / this)
ngl this was a real mess to make kiiind of work and it ultimately looks a
bit hacked together. the basic idea is, we are creating the global object in
the generated code rather than in the runtime initialization, which allows
us to know the imported indexes of all functions. unfortunately, this kind
of meant shoehorning some stuff into NotWasm i probably wouldn't have added
otherwise (GetPrimFunc)