Do not unescape strings
An irritating attribute of Ressa was that it did not process
escaped strings. So, we would receive the JavaScript token "\n"
as the Rust string "\"\\n\"".

Never in my life have I seen a parser that does that.

Anyway, we don't need all the code that I've deleted any more.