fixed notwasm typechecker arity mismatch bug
Before the fix, the notwasm typechecker would report an error
message for an arity mismatch by displaying the difference
between the expected and actual number of arguments.

Because it subtracted these 2 numbers to get their difference,
it could be negative. However, these 2 numbers are of type
`usize` and therefore panic when they underflow.

Arity mismatches now display a real error message instead of
crashing the compiler.