Insert `return undefined` extremely generously
There is significant room for improvement here, though most of it would be hard
to be more precise about (does a statement throw; does a loop exit) at least
the labeled block could be made to return true when returning before a break.

There could also be better testing here. There's no jankyscript parsing,
so it'd be a pain to write proper unit tests. When a function returns in
integration tests, obviously we don't see the inserted return undefined as
a user.

So we know it is inserting returns, and I've manually checked that it's NOT
inserting returns in certain reasonable spots, but that's all.