Support operators coercing FROM any instead of TO
The support for coercing to was extraneous because it's already supported
by every expression. The support for coercing from was necessary to avoid
implementing any versions of operators with known types, like !, &, and more
controversially, <

This removes the weight that prefers "any case." i don't know why we would
prefer that when the coercions sure to be otherwise introduced already
anti-prefer that case. If the types are already any, then we can't coerce
anyway (unless we have, eg, !, in which case we have no choice)